refactor: ignore headers directory
