--- src/readme.md
@@ -1,19 +1,0 @@
-Kin-lang - Interpreter
-===================
-
-Here is where You'll find implementation of our ```Kin-lang's``` ```Tree-Walk Interpreter```.
-
-How it Works
-------------
-
-source-codes -> lexer(): array of tokens -> parser(): abstract binary tree *=ast=* -> interpreter(): execute codes by walking our *ast*
-
-Contribute
-----------
-
-Please refer to [Contributing.md](../contributing.md).
-
-Licese
-------
-
-This project is under [Apache License 2.0 LICENSE](https://github.com/kin-lang/kin/blob/main/LICENSE).
